llm-ros2-agent/requirements.txt
google-generativeai
pyttsx3